chalice/requirements.txt
chalice
line-bot-sdk